Added support for Scheme number semantics.
Adds support for the numeric tower and Scheme number
semantics. This includes complex numbers, which is
newly supported as of this commit.

Support comes from a scheme numbers library by Danny Yoo.
It differs from the eslint style pretty drastically, so the config file was
updated to ignore it for now. The eslint version had to be bumped up to make
this easier.

Minor refactoring and bug fixes are included.

Scheme numbers are off by default.