Fix unnecessary 'links' file copying.
Fixes a bug where the private/interop module was copied to the build
directory if a source file required racketscript/interop.

The racketscript/private/interop module includes the #%js-ffi
function which raises an error if used in Racket. The compiler
overrides this binding with js-ffi calls, so the module
never actually needs to be copied to the build directory.